move aggregate body validator to validation crate
