+++ base_layer/core/src/validation/helpers.rs
@@ -31,6 +31,7 @@ use tari_crypto::{
 };
 use tari_script::TariScript;
 
+use super::aggregated_body::InternalConsistencyAggregateBodyValidator;
 use crate::{
     blocks::{Block, BlockHeader, BlockHeaderValidationError, BlockValidationError},
     borsh::SerializedSize,
@@ -236,9 +237,10 @@ pub fn check_accounting_balance(
     let offset = &block.header.total_kernel_offset;
     let script_offset = &block.header.total_script_offset;
     let total_coinbase = rules.calculate_coinbase_and_fees(block.header.height, block.body.kernels());
-    block
-        .body
+    let body_validator = InternalConsistencyAggregateBodyValidator::default();
+    body_validator
         .validate_internal_consistency(
+            &block.body,
             offset,
             script_offset,
             bypass_range_proof_verification,
